채팅방에 관련된 기능을 RoomApiController로 이동 및 API 수정
