+++ src/main/java/com/lhs/chatting/controller/UserApiController.java
@@ -1,7 +1,5 @@
 package com.lhs.chatting.controller;
 
-import java.util.List;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,9 +12,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.lhs.chatting.model.ChangeUserInfoRequest;
-import com.lhs.chatting.model.GetAllMemberResponse;
 import com.lhs.chatting.model.RegisterUserRequest;
-import com.lhs.chatting.model.entity.Member;
 import com.lhs.chatting.model.entity.User;
 import com.lhs.chatting.service.MemberService;
 import com.lhs.chatting.service.UserService;
@@ -46,15 +42,6 @@ public class UserApiController {
     public ResponseEntity<Long> getUserIdByEmail(@RequestParam String email) {
         Long userId = userService.getUserIdByEmail(email);
         return ResponseEntity.ok(userId);
-    }
-
-    @GetMapping("/{userId}/members")
-    public ResponseEntity<GetAllMemberResponse> getMembers(@PathVariable Long userId){
-        List<Member> members = memberService.getMembers(userId);
-        GetAllMemberResponse response = GetAllMemberResponse.builder()
-                .members(members)
-                .build();
-        return ResponseEntity.ok(response);
     }
     
     @GetMapping("/{userId}")
